xbee: reduced code duplication
